NODE-1915: Some fixes
